// Frontend/src/api/generated/model/communicationDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { CommunicationOriginDto } from './communicationOriginDto';
import type { CommunicationResponsibleDto } from './communicationResponsibleDto';
import type { CommunicationCategoryDto } from './communicationCategoryDto';
import type { CommunicationStatusDto } from './communicationStatusDto';
import type { CommunicationFormatDto } from './communicationFormatDto';
import type { CommunicationRelatedDto } from './communicationRelatedDto';

export interface CommunicationDto {
  id?: string;
  origin?: CommunicationOriginDto;
  date?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  responsibles?: CommunicationResponsibleDto[] | null;
  category?: CommunicationCategoryDto;
  status?: CommunicationStatusDto;
  format?: CommunicationFormatDto;
  /** @nullable */
  relatedId?: string | null;
  related?: CommunicationRelatedDto;
}