// Frontend/src/api/generated/model/communicationMasterDataDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { CommunicationYearDto } from './communicationYearDto';
import type { CommunicationCategoryDto } from './communicationCategoryDto';
import type { CommunicationStatusDto } from './communicationStatusDto';
import type { CommunicationOriginDto } from './communicationOriginDto';
import type { CommunicationResponsibleDto } from './communicationResponsibleDto';
import type { CommunicationFormatDto } from './communicationFormatDto';

export interface CommunicationMasterDataDto {
  /** @nullable */
  years?: CommunicationYearDto[] | null;
  /** @nullable */
  categories?: CommunicationCategoryDto[] | null;
  /** @nullable */
  statuses?: CommunicationStatusDto[] | null;
  /** @nullable */
  origins?: CommunicationOriginDto[] | null;
  /** @nullable */
  responsibles?: CommunicationResponsibleDto[] | null;
  /** @nullable */
  formats?: CommunicationFormatDto[] | null;
}