// Frontend/src/api/generated/model/updateCommunicationRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */

export interface UpdateCommunicationRequest {
  categoryId?: number;
  originId?: number;
  /** @nullable */
  formatId?: number | null;
  date?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  description?: string | null;
  statusId?: number;
  /** @nullable */
  responsibleIds?: string[] | null;
  /** @nullable */
  relatedId?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  filePath?: string | null;
}